# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{DEV} Pedia",
        "title": "{DEV} Pedia API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:9001",
    "basePath": "/",
    "paths": {
        "/api/articles": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List articles with optional filtering and pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "List articles",
                "parameters": [
                    {
                        "enum": [
                            "doc",
                            "git"
                        ],
                        "type": "string",
                        "description": "Source type",
                        "name": "source_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Parent path filter",
                        "name": "parent_path",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new article with the given data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Create a new article",
                "parameters": [
                    {
                        "description": "Article data",
                        "name": "article",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.CreateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.Article"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/articles/by-path": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get an article by its full path",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Get article by path",
                "parameters": [
                    {
                        "enum": [
                            "doc",
                            "git"
                        ],
                        "type": "string",
                        "description": "Source type",
                        "name": "source_type",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Full path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleWithTags"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/articles/search": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Search articles by title or path for autocomplete",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "articles"
                ],
                "summary": "Search articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Result limit (max 50)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ArticleSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/graph/article/{source_type}/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a subgraph centered on a specific article (N-hop neighborhood)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "graph"
                ],
                "summary": "Get article neighborhood graph",
                "parameters": [
                    {
                        "enum": [
                            "doc",
                            "git"
                        ],
                        "type": "string",
                        "description": "Source type",
                        "name": "source_type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 2,
                        "description": "Depth of neighborhood (1-5)",
                        "name": "depth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.GetGraphResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "fceek_dev-pedia_backend_internal_models.Article": {
            "type": "object",
            "properties": {
                "classification_level": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "full_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleMetadata"
                },
                "parent_path": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "source_type": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleSourceType"
                },
                "status": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleStatus"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.ArticleListResponse": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleWithTags"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "fceek_dev-pedia_backend_internal_models.ArticleMetadata": {
            "type": "object",
            "additionalProperties": true
        },
...
        "fceek_dev-pedia_backend_internal_models.ArticleWithTags": {
            "type": "object",
            "properties": {
                "classification_level": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "full_path": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleMetadata"
                },
                "parent_path": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "source_type": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleSourceType"
                },
                "status": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleStatus"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleTag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "fceek_dev-pedia_backend_internal_models.CreateArticleRequest": {
            "type": "object",
            "required": [
                "classification_level",
                "content",
                "full_path",
                "source_type",
                "title"
            ],
            "properties": {
                "classification_level": {
                    "type": "integer",
                    "maximum": 5,
                    "minimum": 1
                },
                "content": {
                    "type": "string"
                },
                "full_path": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleMetadata"
                },
                "parent_path": {
                    "type": "string"
                },
                "secrets": {
                    "description": "New: secrets to create with article",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.CreateContentSecretRequest"
                    }
                },
                "slug": {
                    "type": "string",
                    "maxLength": 100
                },
                "source_type": {
                    "enum": [
                        "doc",
                        "git"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleSourceType"
                        }
                    ]
                },
                "status": {
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/fceek_dev-pedia_backend_internal_models.ArticleStatus"
                        }
                    ]
                },
                "tag_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string",
                    "maxLength": 200
                }
            }
        },
...
        "internal_handlers.ArticleSearchResponse": {
            "type": "object",
            "properties": {
                "suggestions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_handlers.ArticleSuggestion"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer token for regular authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GodToken": {
            "description": "God token for bootstrap operations",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}